// src/client/types.ts
export type VNode = {
  type: string | Component<any> | typeof Fragment;
  props: Props;
  key: string | number | null;
};

export type Props = Record<string, any> & { children?: Child | Child[] };
export type Child = VNode | string | number | boolean | null | undefined;
export type Component<P = Props> = (props: P) => VNode | null;

export const Fragment = Symbol("Fragment");